requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
boto3==1.37.18
python-dotenv==1.0.1
requests==2.32.3
pytest==8.3.5
httpx==0.28.1